chore: Move new endpoints to alert configs
