+++ api/src/lib.rs
@@ -41,14 +41,14 @@ pub fn rocket() -> Rocket<Build> {
                 monitors::get_monitor,
                 monitors::delete_monitor,
                 monitors::update_monitor,
-                monitors::associate_alert_configs,
-                monitors::disassociate_alert_config,
                 jobs::get_job,
                 jobs::start_job,
                 jobs::finish_job,
                 api_keys::list_api_keys,
                 api_keys::generate_key,
                 api_keys::revoke_key,
+                alert_config::associate_alert_configs,
+                alert_config::disassociate_alert_config,
                 alert_config::list_alert_configs,
                 alert_config::create_alert_config,
                 alert_config::get_alert_config,
